requirements.txt
crewai
python-dotenv
streamlit
pypdf
crewai-tools
langchain
litellm
sentence-transformers
torch
transformers